Creating Exception handling - delete

+++ src/main/java/com/GabrielLopes/webService/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.GabrielLopes.webService.entities.User;
 import com.GabrielLopes.webService.repositories.UserRepository;
+import com.GabrielLopes.webService.services.exceptions.DatabaseException;
 import com.GabrielLopes.webService.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +33,15 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id))
+				throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update(Long id, User obj) {
